TCOMP-69: Add the missing parameters for new Salesforce components https://jira.talendforge.org/browse/TCOMP-69 Note:Refactor the salesforce output and bulk properties
